Create solution.java
The solution.java file has two methods:

i) aboveBelow: This Method or function takes two parameters as input. The first parameter is a List of integer and the second parameter is an integer which is used to compare with the list. It returns an map object with the keys "above" and "below" with the corresponding count of integers from the list that are above or below the comparison value.

ii) stringRotation: This Method or function takes two parameters as input. The first parameter is the string and the second parameter is an integer i.e. the number of times the string needs to rotated right side. It returns an new string, rotating the characters in the original string to the right by the rotation amount and have the overflow appear at the beginning